fix(projects-service): change Milestone.sort_order to i32 to match INTEGER column
sqlx 0.8 decodes INT4 (PostgreSQL INTEGER) strictly into i32, not i64.
The FromRow derive was trying to decode sort_order into i64, causing all
milestone reads (create response, list) to fail with a type mismatch error.

Change Milestone.sort_order: i64 -> i32 and update update_milestone to
map the incoming i64 request value to i32 before use.

+++ projects-service/src/lib/models.rs
@@ -61,7 +61,7 @@ pub struct Milestone {
     pub description: Option<String>,
     pub due_date: Option<String>,
     pub status: String,
-    pub sort_order: i64,
+    pub sort_order: i32,
     pub created_at: String,
     pub updated_at: String,
 }
